+++ pytorch_lightning/loops/epoch/training_epoch_loop.py
@@ -39,8 +39,7 @@ class TrainingEpochLoop(loops.Loop):
         self.min_steps: int = min_steps
         self.max_steps: int = max_steps
         self.global_step: int = 0
-        # the total batch index across all epochs
-        self.total_batch_idx: int = 0
+        # manually tracking which is the last batch is necessary for iterable dataset support
         self.is_last_batch: Optional[bool] = None
         self.batch_progress = Progress()
         self.scheduler_progress = SchedulerProgress()
@@ -52,6 +51,13 @@ class TrainingEpochLoop(loops.Loop):
         self._dataloader_idx: Optional[int] = None
         self._warning_cache: WarningCache = WarningCache()
         self._epoch_output: Optional[List[List[STEP_OUTPUT]]] = None
+
+    @property
+    def total_batch_idx(self) -> int:
+        """Returns the current batch index (across epochs)"""
+        # use `ready` instead of `completed` in case this is accessed after `completed` has been increased
+        # but before the next `ready` increase
+        return self.batch_progress.total.ready - 1
 
     @property
     def batch_idx(self) -> int:
@@ -176,13 +182,8 @@ class TrainingEpochLoop(loops.Loop):
         # update plateau LR scheduler after metrics are logged
         self.update_lr_schedulers("step", update_plateau_schedulers=True)
 
-        self.total_batch_idx += 1
-
         # progress global step according to grads progress
         self._increment_accumulated_grad_global_step()
-
-        if self.done:
-            raise StopIteration
 
     def on_run_end(self) -> List[List[STEP_OUTPUT]]:
         """Calls the on_epoch_end hook.
@@ -351,7 +352,7 @@ class TrainingEpochLoop(loops.Loop):
         """Increments global step according to grads progress"""
         if not self._should_accumulate():
             self.global_step = self.trainer.accelerator.update_global_step(
-                self.total_batch_idx, self.trainer.global_step
+                self.batch_progress.current.ready, self.trainer.global_step
             )
 
     def _should_check_val_fx(self, batch_idx: int, is_last_batch: bool) -> bool: